+++ spec/unit/installer_spec.rb
@@ -93,6 +93,51 @@ module Pod
         @installer.install!
       end
 
+      it 'runs source provider hooks before analyzing' do
+        config.skip_repo_update = true
+        @installer.unstub(:resolve_dependencies)
+        @installer.stubs(:validate_build_configurations)
+        @installer.stubs(:prepare_for_legacy_compatibility)
+        @installer.stubs(:clean_sandbox)
+        def @installer.run_source_provider_hooks
+          @hook_called = true
+        end
+        def @installer.analyze(*)
+          @hook_called.should.be.true
+        end
+        @installer.install!
+      end
+
+      it 'includes sources from source provider plugins' do
+        plugin_name = 'test-plugin'
+        Pod::HooksManager.register(plugin_name, :source_provider) do |context, options|
+          source_url = options['sources'].first
+          return unless source_url
+          source = Pod::Source.new(source_url)
+          context.add_source(source)
+        end
+
+        test_source_name = 'https://github.com/artsy/Specs.git'
+        plugins_hash = Installer::DEFAULT_PLUGINS.merge(plugin_name => { 'sources' => [test_source_name] })
+        @installer.podfile.stubs(:plugins).returns(plugins_hash)
+        @installer.unstub(:resolve_dependencies)
+        @installer.stubs(:validate_build_configurations)
+        @installer.stubs(:prepare_for_legacy_compatibility)
+        @installer.stubs(:clean_sandbox)
+        @installer.stubs(:ensure_plugins_are_installed!)
+        @installer.stubs(:analyze)
+        config.skip_repo_update = true
+
+        analyzer = Installer::Analyzer.new(config.sandbox, @installer.podfile, @installer.lockfile)
+        analyzer.stubs(:analyze)
+        @installer.stubs(:create_analyzer).returns(analyzer)
+        @installer.install!
+
+        source = Pod::Source.new(test_source_name)
+        names = analyzer.sources.map(&:name)
+        names.should.include(source.name)
+      end
+
       it 'integrates the user targets if the corresponding config is set' do
         config.integrate_targets = true
         @installer.expects(:integrate_user_project)
@@ -724,6 +769,14 @@ module Pod
         Installer::PostInstallHooksContext.expects(:generate).returns(context)
         HooksManager.expects(:run).with(:post_install, context, Installer::DEFAULT_PLUGINS)
         @installer.send(:run_plugins_post_install_hooks)
+      end
+
+      it 'runs plugins source provider hook' do
+        context = stub
+        context.stubs(:sources).returns([])
+        Installer::SourceProviderHooksContext.expects(:generate).returns(context)
+        HooksManager.expects(:run).with(:source_provider, context, Installer::DEFAULT_PLUGINS)
+        @installer.send(:run_source_provider_hooks)
       end
 
       it 'only runs the podfile-specified hooks' do